Add child element order for CNA
We've been lucky thus far in that apparently nobody has ever modified a
CNA once it's been created, so the element order imposed by CNA.bld was
never challenged.

Commit b27a1d6e0f1bad80b7b4aad8b359f123ffe8294d changed all that by
replacing UseNextAvailableHighSlotID with UseNextAvailableSlotID on the
fly before PUT.  This would have worked fine if CNA had ever had a child
element order defined.

It didn't.

So we define it via this change set.

We also fix a latent typo in the spelling of the
VirtualStationInterfaceTypeID field, which apparently nobody has ever
used, or they would have broken.

Closes-Bug: #1688544
Change-Id: I5ac82d55bdab4e681b1ba0ba53992c795248700d