Fix Ascii Decode Error in Open Vterm
When running vterm.open_remotable_vnc_vterm on a system that is
configured for a language that contains double byte characters,
when the mkvterm commmand fails and returns these double byte
characters, either the vterm module will blow up in the
_run_mkvterm_cmd method or the caller will blow up when trying
to format the exception.  The reason for this is that the
subprocess communicate method will return the stdout/stderr as
a byte string, so any code then that implicitly tries to convert
it to a unicde string (through the string formatter) will blow
up since it isn't going to use UTF-8 for the implicit convert.

Rather than just doing this conversion from byte-string to a
unicode string locally in _run_mkvterm for the exception and
forcing the caller to always do an explict conversion, making
a change to convert the stdout/stderr to a unicode string in the
_run_proc method so that everyone above that method is good
with it already being a unicode string.

Change-Id: I71268e3f05e5d9afeb08a711bbdaa6c9625e88d7