Make power_off honor PowerOffOpts with operation
The recent refactor of the pypowervm.tasks.power module [1] left some
code paths unaccounted for: specifically where power_off was called with
add_params as a PowerOffOpts with an actual operation already set, as
opposed to just using the PowerOffOpts for e.g. restart or immediate.

The behavior in such scenarios was undefined, but definitely not what
would reasonably be expected.

This change set causes such invocations to be executed with the
PowerOffOpts as is, ignoring the restart and force_immediate parameters.

[1] https://github.com/powervm/pypowervm/commit/5bd2a28352db55dc564c79aaa285c61d0c1fe135

Change-Id: Ib954249f09eab4f646b88342595f9e8c7e161c85