Don't fail live SRR toggle if capable
It used to be the case that an LPAR had to be shut down in order to
toggle the Simplified Remote Restart (SRR) capability.  The platform
changed that and added a new capability bit to the ManagedSystem,
brought into pypowervm via [1].  However, the CapabilitiesValidator
utility used by LPARBuilder during a resize operation still had a check
that would cause the resize to fail if it included a change to the SRR
capability bit for a live LPAR.  With this change, we update that
condition to take into account the new "dynamic SRR" capability, only
failing validation if that capability is False.

[1] https://github.com/powervm/pypowervm/commit/97f15c2e646e4d5273f1dd8ddb684430819a09b6

Change-Id: I4c3dc62a1db2a6f9844863b7a7a4e2ea68ee89f4