Return 1 if file does not exist for partition id
The call of my_partition_id assumes that the file
/proc/ppc64/lparcfg will always exist.  This may not be true
for certain distros.  However, if it doesn't exist, it can
be assumed to be 1.

The outlier case is if the adapter is 'remote'.  However, since
the adapter is not passed in, it can't be determined if the
connection is remote.  In which case, returning a default of 1
seems more appropriate than the error the user would get today.

Change-Id: I517c225774a84d1054efa30b693dfb8320c44c3b