Open vSwitch integration into CNA
The CNA will now accept the Open vSwitch attributes.  However,
these attributes are currently only valid if the CNA is hosted on
the NovaLink partition, and is a trunk adapter.  Per our precedent,
we let the REST API enforce those attributes restrictions as some
may be lifted at a later point.

Change-Id: I50b5f45ef338bb2ed142e95482b9d1f06966a8dc