Add PPT ratio support
This change adds support for viewing the system and LPAR attributes
related to PPT ratio and enables the setting of the PPT ratio value
on LPARs.

The partition page table ratio represents the ratio of a VM's maximum
memory to the size of its partition page table. The PPT is used by the
platform to maintain the translation of the VM's physical to virtual
memory addresses during mobility operations (i.e. LPM).

The ppt_ratio is represented as 1:2^N where accepted values for N range
from 6 to 12. If the ratio is 1:4096 for a VM with 64 GB of maximum
memory, it would have a PPT of 16 MB (64 GB / 4096 = 16 MB).

The PPT ratio will be passed into the LPAR builder in the list of
attributes as a string representing the ratio (e.g. '1:4096'). The LPAR
builder converts the ratio to the REST accepted format.

Change-Id: I147767ce9260c5275260117b797c60ccdc20c694