Accept 0 for capacity, max_capacity values for ethlp/vnic bld
This fixes the issue of setting capacities if they are 0
when building a logical port or a VNIC backing device
wrapper.

It does not handle any valid range validation such as
verifying the capacity is smaller or equal to max_capacity,
that the capacities match granularity, etc. That is left
to the REST layer.

Change-Id: Ib7d7927abdb50c1994d92fc58e96f4a3b6dda390