Add support to allow performance data collection for LPAR
This change adds support for setting the LPAR attribute
that governs performance data collection.
Hypervisor will collect performance metric for the lpar
once the lpar metric collection is enabled.

The enable lpar metric wiil be passed into the LPAR builder
in the list of attributes as a True or False.
Change-Id: Ib78158e60b48889a34cda2a6e345f73f4aed53e8