Add partition assignment info to System IOSlot
REST recently added IOSlot assignment information to the
ManagedSystem.AssociatedIOConfiguration.IOSlots.IOSlot elements in the
form of fields PartitionID, PartitionUUID, PartitionName, and
PartitionType.  This change accomodates the new fields on
managed_system.IOSlot as part_id, part_uuid, part_name, and part_type,
respectively.

With the REST change, a capability field, IOSlotOwnerAssignmentCapable,
was added (so that, when the slot assignment data is absent, we can tell
the difference between "slot is unassigned" and "running an old REST
server").  This change adds the new capability bit as
'ioslot_owner_assignment_capable'.

Change-Id: If630e8d5cb0b713847ac37418f96150f2e823feb
Partial-Implements: bp/pci-passthrough