Add default PPT ratio property
The reason we're adding the default for the PPT ratio where
we don't do this type of thing traditionally is because we
need a way to determine if the PPT ratio has been set by
the user.

A good way to do this is to check if it's different than the
default. With this we can determine if we should migrate
LPARs to PPT ratio capable hosts, or incapable hosts where
the LPAR would lose its set PPT ratio value.

Change-Id: I416b6f397554d7bafafdf93fb9ceaa03a40ed6a9