Session deepcopy (correctly) raises in py36
Deep-copying a Session isn't supported.  It was being done in a unit
test for an edge case on garbage collection, but really only in py3.
py36 closed the gap that was incorrectly open in <=py35 and properly
made this raise TypeError again.  So we have to be a little more finicky
about that code path where we expect it to go through.

Change-Id: Id11ecfac7f9af2196b80d2b808364f433a85b74e
Closes-Bug: #1722413