Don't preserve backing port-less VFC mappings
derive_npiv_map attempts to check for existing mappings to the
v_port_wwpns and p_port_wwpns pairs passed in. In the case where
there's an existing mapping for the v_port_wwpns we're looking
for but the mapping doesn't have a backing port, ignore it.

This is the same as passing in preserve=False but *only* for this
mapping.

Closes-Bug: #1722412
Change-Id: I9f3ca1e75229940ccdbe9e8c85e512734149daba