Set vdtype to LogicalVolume for VDisk.bld_ref
To create a VSCSI mapping against an existing Logical Volume, one
expects to be able to populate the storage piece of the new mapping with
a LV reference created via VDisk.bld_ref. However, sending this down to
REST results in a 500 from a NullPointerException because the
VirtualDiskType isn't set. This is because, before FileIO and RBD were
introduced, VirtualDisk was always a LogicalVolume, and there was no
such thing as VirtualDiskType.

With this change, we set the VirtualDiskType to LogicalVolume for
VDisk.bld_ref.

We also introduce an alias for VDisk called LV to make it explicit that
this is always a Logical Volume now.

Change-Id: I23170b8c6c24e119dda88901c47245f118b9e1e8
Closes-Bug: #1731507