Let pbr handle files and changes
MANIFEST.in is removed since pbr will automatically generate this
file based on the files that git is tracking.

VERSION.rst has not been updated in a very long time. Rather than
try to bring that up to date, it is simply removed.

The development status classifier in setup.cfg is also updated to
indicate that this is no longer alpha, which it has not been for a
very long time.

Change-Id: Ifcb078e17443dcbe2776ac7ffe236d9f3a272a7c