fix oslo deb names
Canonical switched from naming their oslo debs python-oslo-* to
python-oslo.* a long time ago, but allowed the old names to work as
well for a while. That is ending, and we still have a couple
references to the old naming format that we need to fix. This is
necessary to be able to use pypowervm with OpenStack's Queens
release.

Change-Id: I448ba1ecfa0551dd4b9131838832c5c758ff1b14