Support setting storage dev tag during bld[_ref]
External agents may wish to set an arbitrary string tag on a storage
device to allow subsequent correlation of the device with some metadata
representation of it.

This change set adds a `tag` kwarg to the bld and bld_ref methods of PV,
VDisk, RBD, FileIO, and LU.

The value is passed through as the Tag property in the REST payload,
which the platform will recognize and use to tag the device when it is
mapped.

Change-Id: I3095ed21fc8dc08488684d3248aecf5c1a4df008