Support setting PV device tag during bld
External agents may wish to set an arbitrary string tag on a physical
volume to allow subsequent correlation of the PV with some
metadata representation of it.  This change set adds a `device_tag`
kwarg to PV.bld.  The value is passed through as the VolumeID property
in the REST payload, which the platform will recognize and use to tag
the PV when it is mapped.

Change-Id: I3095ed21fc8dc08488684d3248aecf5c1a4df008