Cache physical WWPNs
Since they're expensive to retrieve, add the ability to cache physical
WWPNs gleaned from the VIOSes.  This is done via a new force_refresh
kwarg to get_physical_wwpns.  When False, caching is used.  The value is
True for backward compatibility.

NOTE: if you hot plug adapters, the cache doesn't get refreshed
automatically - you have to invoke the method with force_refresh=True
(or restart the process).

At some point in the future, we may be able to consume an event from
REST that will prompt us to force refresh the cache automatically in
this scenario.

Change-Id: I9370cf1f1ee7bf5edfa8f5f7ef59dbf8000547da