Add affinity check support as part of LPM
This change adds support for viewing the system capability
related to VM affinity score check as part of migration and
enables setting of the affinity check metadata on VMs as
part of deploy.

The value for this metadata is represented as true or false.
If set to true, that means affinity score check on the
destination host needs to be performed for the VM as part of
LPM operation by passing a extra flag(Affinity) to the
job parameter.

The default behaviour is not to check for the score on the
destination host, and this is not a mandatory parameter to be
passed to the LPM job.

Change-Id: Icac0579aae51c48ba744909bec02b2bbda8348cc