More tox cleanup
This attempts to cleanup a few things in our tox.ini as follows:
1) Removes the "-U" argument in the install command. This is not
necessarily required and we should leave argument selection up to
the user. This is also consistent with other projects.
2) Sets basepython for venv to use python 3. Python 2 will be EOL
soon so we need to start treating things as py3 first.
3) Removes several things from flake8's exclude list that we don't
have (e.g. docs). We may be able to trim this further, but I tried
to be somewhat conservative.

Change-Id: I0c8f77d35ddf9a4793424d4f442ab688f7b5d94b