Fix python 3 vterm error
To retrieve the next item from the iterator,

  - Python 2.7 provides 2 ways:
    1) iter.next()
    2) next(iter)     --- implicitly calls iter.next()

  - Python 3.5 provides 2 ways:
    1) iter.__next__()
    2) next(iter)     --- implicitly calls iter.__next__()

This replaces an iter.next() call, which would be py2-specific,
with a next(iter) call that will work for both py2 and py3.

UT was not hitting this because it was not testing the path with
this call. UT is herein updated to cover this path.

Change-Id: If228397866a2f1cb3cdb77087ba9e645f2509022