Disable secure boot for IBMi deploys
The secure boot attribute was being added to the lpar wrapper object
as part of VM deploy, irrespective of the OS type. This doesn't work
for IBMi, which currently blocks secure boot at the schema level.
This fix adds validation for IBMi (caller should not request secure
boot) and avoids adding the attribute to the lpar wrapper.

Change-Id: I1a786265c38b11fe44936ec628bcfdd26d5d20be
Closes-Bug: #1805610