Add the device ID parameter to discover hdisk
LUA recovery will sometimes find multiple existing hdisks at the
same ITLs. Without the device ID, LUA recovery will return a bad
status.

This change makes it so that discover_hdisk accepts a deviceID
parameter and appends it to the input XML. This enables LUA recovery
to identify the correct hdisk when there are multiple on the same
ITL.

For most volume backed storage planes such as SVC, the device ID
is the base64 encoded value of the pg83 identifier.

Change-Id: I6788592ad228f8cf3c1ee5775cd8799d58188e58