Fix schema issue related to RoCE SR-IOV adapters
Update the wrapper code to handle schema changes related
to RoCE SR-IOV adapter response. The reponse for this
kind of adapter doesn't follow the same format as other
SR-IOV adapters which leads a case where the physical ports
information could not be processed. The proposed fix here
address this issue.

Change-Id: If10bafa0e5cd8022a4d8d366fbaf8b15d2cce08b