Handle UnicodeDecodeError during slot_map fetch.
In case of an upgraded env, where Novalink is
at the latest level and the python version
got bumped up from 2 to 3 as part of the upgrade,
for the VMs deployed with older version of python
the slot_map data fetch call will fail with
unicode error. This is due to the changes we
made in py3 to encrypt/decrypt the data using base64.
This change will handle the issue with slot_map
fetch call as the data type for VMs deployed on
py2 will be different and the py3 change will
not work in such cases.

Change-Id: I03d03187e307d8be78fc3ec9d0af01be42d5eac4